Ajout sécurité controller

+++ src/main/java/com/myaudiolibrary/web/controller/AlbumController.java
@@ -10,6 +10,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
+import javax.persistence.EntityNotFoundException;
+import java.util.Optional;
+
 
 @RestController
 @RequestMapping(value = "albums")
@@ -21,6 +24,10 @@ public class AlbumController {
     @RequestMapping(method = RequestMethod.POST,consumes = MediaType.APPLICATION_JSON_VALUE,produces =
             MediaType.APPLICATION_JSON_VALUE)
     public Album ajoutAlbum(@RequestBody Album album){
+
+        if(album.getTitle().isEmpty() || album.getTitle().isBlank())
+            throw new IllegalArgumentException("Le nom de l'album n'est pas renseigné");
+
         Album result = albumRepository.save(album);
         return result;
     }
@@ -29,6 +36,16 @@ public class AlbumController {
             MediaType.APPLICATION_JSON_VALUE,value = "/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void ajoutAlbum(@PathVariable("id") Integer id){
+
+        if (id<1)
+            throw new IllegalArgumentException("L'id de l'album doit être égal ou supérieur à 1" );
+
+        Optional<Album> album = albumRepository.findById(id);
+
+        if(album.isEmpty())
+            throw new EntityNotFoundException("L'album avec l'id " + id + " n'as pas été trouvé");
+
+
         albumRepository.deleteById(id);
     }
 }
